extracted body container
